backend/requirements.txt
fastapi~=0.78.0
pydantic~=1.9.1
SQLAlchemy~=1.4.37
python-dotenv~=0.20.0
alembic~=1.8.0
passlib~=1.7.4
sqlmodel==0.0.4
asyncpg==0.24.0
uvicorn==0.17.6
pymysql==0.7.11
sqlalchemy_utils~=0.32.2
colour~=0.1.5